Delete integration tests code
